move all business logic to presenter
