+++ app/src/main/java/com/dicoding/picodiploma/sejiwaproject/features/match/detail/DetailMatchActivity.kt
@@ -1,6 +1,5 @@
 package com.dicoding.picodiploma.sejiwaproject.features.match.detail
 
-import android.database.sqlite.SQLiteConstraintException
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
@@ -11,15 +10,10 @@ import com.dicoding.picodiploma.sejiwaproject.R
 import com.dicoding.picodiploma.sejiwaproject.commons.api.ApiRepository
 import com.dicoding.picodiploma.sejiwaproject.commons.utils.invisible
 import com.dicoding.picodiploma.sejiwaproject.commons.utils.visible
-import com.dicoding.picodiploma.sejiwaproject.db.Favorite
 import com.dicoding.picodiploma.sejiwaproject.db.database
 import com.dicoding.picodiploma.sejiwaproject.features.match.detail.model.DetailMatch
 import com.google.gson.Gson
 import kotlinx.android.synthetic.main.activity_detail_match.*
-import org.jetbrains.anko.db.classParser
-import org.jetbrains.anko.db.delete
-import org.jetbrains.anko.db.insert
-import org.jetbrains.anko.db.select
 
 class DetailMatchActivity : AppCompatActivity(),
     DetailMatchView {
@@ -36,16 +30,17 @@ class DetailMatchActivity : AppCompatActivity(),
 
         matchDetail = intent.getStringExtra(EXTRA_ID) as String
 
-        favoriteState()
         val request = ApiRepository()
         val gson = Gson()
         presenter =
             DetailMatchPresenter(
                 this,
                 request,
-                gson
+                gson,
+                database
             )
         presenter.getDetailMatch(matchDetail)
+        presenter.favoriteState(matchDetail)
 
     }
 
@@ -61,16 +56,9 @@ class DetailMatchActivity : AppCompatActivity(),
         progress_bar.invisible()
     }
 
-    private fun favoriteState() {
-        database.use {
-            val result = select(Favorite.TABLE_FAVORITE)
-                .whereArgs(
-                    "(MATCH_ID = {id})",
-                    "id" to matchDetail
-                )
-            val favorite = result.parseList(classParser<Favorite>())
-            if (favorite.isNotEmpty()) isFavorite = true
-        }
+    override fun favoriteState(isFavorite: Boolean) {
+        this.isFavorite = isFavorite
+
     }
 
     override fun matchReady(detailMatch: DetailMatch) {
@@ -133,7 +121,11 @@ class DetailMatchActivity : AppCompatActivity(),
                 true
             }
             R.id.add_to_favorite -> {
-                if (isFavorite) removeFromFavorite() else addToFavorite()
+                if (isFavorite){
+                    presenter.removeFromFavorite(matchDetail)
+                } else {
+                    presenter.addToFavorite(match)
+                }
 
                 isFavorite = !isFavorite
                 setFavorite()
@@ -143,36 +135,6 @@ class DetailMatchActivity : AppCompatActivity(),
 
             else -> super.onOptionsItemSelected(item)
 
-        }
-    }
-
-    private fun addToFavorite() {
-        try {
-            database.use {
-                insert(
-                    Favorite.TABLE_FAVORITE,
-                    Favorite.MATCH_ID to match.matchId,
-                    Favorite.TEAM_HOME to match.teamHome,
-                    Favorite.TEAM_AWAY to match.teamAway,
-                    Favorite.DATE_MATCH to match.dateMatch,
-                    Favorite.BADGE_HOME to match.badgeHome,
-                    Favorite.BADGE_AWAY to match.badgeAway
-                )
-            }
-        } catch (e: SQLiteConstraintException) {
-
-        }
-    }
-
-    private fun removeFromFavorite() {
-        try {
-            database.use {
-                delete(
-                    Favorite.TABLE_FAVORITE, "(MATCH_ID = {id})",
-                    "id" to matchDetail
-                )
-            }
-        } catch (e: SQLiteConstraintException) {
         }
     }
